SwipeRefreshLayoutをViewPagerの外側に移動
