+++ app/src/main/java/jp/co/aizu_student/weatherhacks/views/activities/MainActivity.java
@@ -5,8 +5,10 @@ import android.content.Intent;
 import android.content.SharedPreferences;
 import android.databinding.DataBindingUtil;
 import android.os.Bundle;
+import android.os.Handler;
 import android.support.design.widget.TabLayout;
 import android.support.v4.view.ViewPager;
+import android.support.v4.widget.SwipeRefreshLayout;
 import android.support.v7.widget.Toolbar;
 
 import javax.inject.Inject;
@@ -18,7 +20,8 @@ import jp.co.aizu_student.weatherhacks.helpers.WeatherHacksApiHelper;
 import jp.co.aizu_student.weatherhacks.models.Location;
 import jp.co.aizu_student.weatherhacks.adapter.MyPagerAdapter;
 
-public class MainActivity extends BaseActivity implements ViewPager.OnPageChangeListener {
+public class MainActivity extends BaseActivity
+        implements ViewPager.OnPageChangeListener, SwipeRefreshLayout.OnRefreshListener {
 
     /** リクエストコード */
     public static final int REQUEST_CODE = 1;
@@ -95,6 +98,7 @@ public class MainActivity extends BaseActivity implements ViewPager.OnPageChange
         binding.tabLayout.setupWithViewPager(binding.viewPager);
         binding.tabLayout.setTabGravity(TabLayout.GRAVITY_FILL);
         binding.tabLayout.setTabMode(TabLayout.MODE_FIXED);
+        binding.setRefreshListener(this);
     }
 
     @Override
@@ -107,5 +111,16 @@ public class MainActivity extends BaseActivity implements ViewPager.OnPageChange
 
     @Override
     public void onPageSelected(int position) {
+    }
+
+    @Override
+    public void onRefresh() {
+        new Handler().postDelayed(() -> {
+            weatherHacksApiHelper.requestWeather(
+                    WeatherHacks.getInstance().getLocationId(),
+                    getSupportFragmentManager()
+            );
+            binding.swipeRefreshLayout.setRefreshing(false);
+        }, 1500);
     }
 }
