+++ app/src/main/res/layout/activity_main.xml
@@ -1,6 +1,12 @@
 <layout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto">
 
+    <data>
+        <variable
+            name="refreshListener"
+            type="android.support.v4.widget.SwipeRefreshLayout.OnRefreshListener" />
+    </data>
+
     <RelativeLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent">
@@ -29,17 +35,24 @@
             android:layout_below="@id/tab_layout"
             android:background="#FFFFFFFF">
 
-            <android.support.v4.view.ViewPager
-                android:id="@+id/view_pager"
+            <android.support.v4.widget.SwipeRefreshLayout
+                android:id="@+id/swipe_refresh_layout"
                 android:layout_width="match_parent"
-                android:layout_height="match_parent" />
+                android:layout_height="wrap_content"
+                android:onSwipeRefresh="@{refreshListener.onRefresh}">
 
-            <View
-                android:id="@+id/shadow"
-                android:layout_width="match_parent"
-                android:layout_height="8dp"
-                android:background="@drawable/bg_shadow" />
+                <android.support.v4.view.ViewPager
+                    android:id="@+id/view_pager"
+                    android:layout_width="match_parent"
+                    android:layout_height="match_parent" />
 
+                <View
+                    android:id="@+id/shadow"
+                    android:layout_width="match_parent"
+                    android:layout_height="8dp"
+                    android:background="@drawable/bg_shadow" />
+
+            </android.support.v4.widget.SwipeRefreshLayout>
         </FrameLayout>
 
     </RelativeLayout>
